added option to output model as csv, with 1 doc per row, started cleaning the console outputs

+++ src/P0_Project/ModelSpecs.java
@@ -42,6 +42,6 @@ public class ModelSpecs {
             outputSimilarity = true;
             similarityOutput = outputDir+similarityOutput;
         }
-        numWordId = Math.toIntExact((long) specs.getOrDefault("numWordId", 3));
+        numWordId = Math.toIntExact((long) specs.getOrDefault("numWordId", (long) 3));
     }
 }
